refactor: setState for radioButton
